+++ Source/CommonExtensions/TransformExtensions.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using JetBrains.Annotations;
 using Omega.Tools;
-using Omega.Tools.Experimental.UtilitiesAggregator;
+using Omega.Package.Internal;
 using UnityEngine;
 
 public static class TransformExtensions
@@ -32,10 +32,20 @@ public static class TransformExtensions
         if (!transform)
             throw new MissingReferenceException(nameof(transform));
 
-        if(result is null)
+        if (result is null)
             throw new ArgumentNullException(nameof(result));
-        
+
         TransformUtilities.GetChildsWithoutChecks(transform, result);
+    }
+
+    public static bool IsChildOf(this Transform self, Transform parent)
+    {
+        if(self is null)
+            throw new NullReferenceException(nameof(self));
+        if (!self)
+            throw new MissingReferenceException(nameof(self));
+
+        return TransformUtilities.IsChildOfWithoutChecks(self, parent);
     }
     
     /// <summary>
@@ -109,5 +119,21 @@ public static class TransformExtensions
         transform.parent = attachTo;
 
         return transform;
+    }
+
+    public static void SetRect(this RectTransform rectTransform, Rect rect)
+    {
+        if (rectTransform is null)
+            throw new NullReferenceException(nameof(rectTransform));
+        if (!rectTransform)
+            throw new MissingReferenceException(nameof(rectTransform));
+
+        RectTransformUtilities.SetRectWithoutChecks(rectTransform, rect);
+    }
+
+    public static bool ToRectTransform(this Transform self, out RectTransform rectTransform)
+    {
+        rectTransform = self as RectTransform;
+        return rectTransform != null;
     }
 }