+++ Source/Routines/Base/WaitAsyncOperationRoutine.cs
@@ -1,14 +1,18 @@
 using System;
 using System.Collections;
+using System.Collections.Generic;
 using UnityEngine;
+using UnityEngine.Networking;
 
 namespace Omega.Routines
 {
-    internal sealed class WaitAsyncOperationRoutine<TAsyncOperation, TResult> : Routine<TResult>, IProgressRoutineProvider
+    internal sealed class WaitAsyncOperationRoutine<TAsyncOperation, TResult> 
+        : Routine<TResult>, IProgressRoutineProvider
         where TAsyncOperation : AsyncOperation
     {
         private TAsyncOperation _asyncOperation;
         private Func<TAsyncOperation, TResult> _resultSelector;
+        private bool? _canBeForcedCompletion;
 
         public WaitAsyncOperationRoutine(TAsyncOperation asyncOperation, Func<TAsyncOperation, TResult> resultSelector)
         {
@@ -25,6 +29,15 @@ namespace Omega.Routines
             SetResult(result);
         }
 
+        protected override void OnForcedComplete()
+        {
+            if (!_canBeForcedCompletion.HasValue)
+                _canBeForcedCompletion = _asyncOperation.CanBeForceComplete();
+            
+            if(!_canBeForcedCompletion.Value)
+                throw new InvalidOperationException("AsyncOperation can not be forced complete");
+        }
+
         public float GetProgress()
         {
             return _asyncOperation.progress;
@@ -34,10 +47,12 @@ namespace Omega.Routines
     internal sealed class WaitAsyncOperationRoutine : Routine, IProgressRoutineProvider
     {
         private AsyncOperation _asyncOperation;
+        private bool? _canBeForcedCompletion;
 
-        public WaitAsyncOperationRoutine(AsyncOperation asyncOperation)
+        public WaitAsyncOperationRoutine(AsyncOperation asyncOperation, bool? canBeForcedCompletion = false)
         {
             _asyncOperation = asyncOperation ?? throw new ArgumentNullException(nameof(asyncOperation));
+            _canBeForcedCompletion = canBeForcedCompletion;
         }
 
         protected override IEnumerator RoutineUpdate()
@@ -49,6 +64,15 @@ namespace Omega.Routines
         public float GetProgress()
         {
             return _asyncOperation.progress;
+        }
+
+        protected override void OnForcedComplete()
+        {
+            if (!_canBeForcedCompletion.HasValue)
+                _canBeForcedCompletion = _asyncOperation.CanBeForceComplete();
+            
+            if(!_canBeForcedCompletion.Value)
+                throw new InvalidOperationException("AsyncOperation can not be forced complete");
         }
     }
 }