+++ Source/Routines/ByEnumerator/ByEnumeratorRoutine.cs
@@ -4,10 +4,13 @@ using JetBrains.Annotations;
 
 namespace Omega.Routines
 {
-    public class ByEnumeratorRoutine : Routine
+    public class ByEnumeratorRoutine : Routine, IProgressRoutineProvider
     {
         [CanBeNull] private IEnumerator _enumerator;
-
+        private float _progress;
+        internal Action OnCancelation;
+        internal Action OnForceComplete;
+            
         internal ByEnumeratorRoutine([CanBeNull] IEnumerator enumerator)
         {
             _enumerator = enumerator;
@@ -27,9 +30,31 @@ namespace Omega.Routines
                 yield return _enumerator.Current;
         }
 
+        internal void SetProgress(float progress)
+        {
+            _progress = progress;
+        }
+
+        protected override void OnCancel()
+        {
+            OnCancelation?.Invoke();
+            base.OnCancel();
+        }
+
+        protected override void OnForcedComplete()
+        {
+            OnForceComplete?.Invoke();
+            base.OnForcedComplete();
+        }
+
         public IEnumerator GetEnumerator()
         {
             return _enumerator;
+        }
+
+        public float GetProgress()
+        {
+            return IsComplete ? 1 : _progress;
         }
     }
 }