+++ Source/Routines/ByEnumerator/ByEnumeratorRoutineWithResult.cs
@@ -4,9 +4,12 @@ using JetBrains.Annotations;
 
 namespace Omega.Routines
 {
-    public class ByEnumeratorRoutine<TResult> : Routine<TResult>
+    public class ByEnumeratorRoutine<TResult> : Routine<TResult>, IProgressRoutineProvider
     {
         [CanBeNull] private IEnumerator _enumerator;
+        internal Action OnCancelation;
+        internal Action OnForceComplete;
+        private float _progress;
 
         internal ByEnumeratorRoutine(Func<RoutineControl<TResult>, IEnumerator> enumeratorGetter)
         {
@@ -21,10 +24,32 @@ namespace Omega.Routines
             while (_enumerator.MoveNext())
                 yield return _enumerator.Current;
         }
+
+        protected override void OnCancel()
+        {
+            OnCancelation?.Invoke();
+            base.OnCancel();
+        }
+
+        protected override void OnForcedComplete()
+        {
+            OnForceComplete?.Invoke();
+            base.OnForcedComplete();
+        }
+
+        internal void SetProgress(float progress)
+        {
+            _progress = progress;
+        }
         
         public IEnumerator GetEnumerator()
         {
             return _enumerator;
+        }
+
+        public float GetProgress()
+        {
+            return IsComplete ? 1 : _progress;
         }
     }
 }