+++ Source/Routines/RoutineWithResult.cs
@@ -1,5 +1,4 @@
 using Omega.Routines.Exceptions;
-using Omega.Tools.Experimental.UtilitiesAggregator;
 
 namespace Omega.Routines
 {